Move responsibility of receiving events from component to service
