+++ src/components/slack/list/slacklist.component.ts
@@ -1,13 +1,15 @@
 import { Component, ChangeDetectorRef, OnInit, OnDestroy } from '@angular/core';
 import { Router } from '@angular/router';
+import { SlackService, DisplaySlackMessageInfo, DisplaySlackReactionInfo } from '../../../services/slack/slack.service';
+
 import {
-    SlackServiceCollection,
     SlackMessage,
-    SlackService,
+    SlackClient,
     SlackReactionAdded,
     SlackReactionRemoved,
-    EmojiService
-} from '../../../services/slack/slack.service';
+} from '../../../services/slack/slack-client';
+
+import { EmojiService } from '../../../services/slack/emoji.service';
 
 import {
     SlackParser,
@@ -24,267 +26,8 @@ import { GlobalEventService } from '../../../services/globalevent.service';
 import { Subscription } from 'rxjs';
 import { Channel, DataStore } from '../../../services/slack/slack.types';
 import { SettingService } from '../../../services/setting.service';
-
-class DisplaySlackReactionInfo {
-    public showReactedUsers = false;
-
-    constructor(public target: DisplaySlackMessageInfo, public rawReaction: string, public reaction: string, public users: string[]) {
-    }
-
-    addUser(user: string) {
-        this.removeUser(user);
-        this.users.push(user);
-    }
-
-    removeUser(user: string) {
-        this.users = this.users.filter(u => u !== user);
-    }
-
-    get count(): number {
-        return this.users.length;
-    }
-
-    get includeMine(): boolean {
-        return !!(this.users.find(u => u === this.target.message.myUserId));
-    }
-
-    get userNames(): string {
-        return this.users.map((userID) => {
-            return this.target.message.dataStore.getUserById(userID).name;
-        }).join(',');
-    }
-}
-
-export class DisplaySlackMessageInfo {
-    edited: boolean = false;
-    image: string = null;
-    reactions: DisplaySlackReactionInfo[] = [];
-
-    constructor(
-        public message: SlackMessage,
-        public parser: SlackParser,
-        public client: SlackService
-    ) {
-    }
-
-    get text(): string {
-        return this.parser.parse(this.message.text, this.message.dataStore);
-    }
-
-    get imageSrc(): string {
-        return `data:${this.message.rawMessage.file.mimetype};base64,${this.image}`;
-    }
-
-    get imageURL(): string {
-        return this.message.rawMessage.file.url_private;
-    }
-
-    get attachments(): Attachment[] {
-        return this.message.rawMessage.attachments
-            ? this.message.rawMessage.attachments
-            : [];
-    }
-
-    get doesReactionExist(): boolean {
-        return this.reactions.length > 0;
-    }
-
-    addReaction(info: SlackReactionAdded) {
-        const reaction = this.parser.parse(`:${info.reaction.reaction}:`, this.message.dataStore);
-        const user = info.reaction.user;
-        const target = this.reactions.find(r => r.reaction === reaction);
-
-        if (target) {
-            target.addUser(user);
-        } else {
-            this.reactions.push(new DisplaySlackReactionInfo(this, info.reaction.reaction, reaction, [user]));
-        }
-    }
-
-    removeReaction(info: SlackReactionRemoved) {
-        const reaction = this.parser.parse(`:${info.reaction.reaction}:`, this.message.dataStore);
-        const target = this.reactions.find(r => r.reaction === reaction);
-
-        if (target) {
-            target.removeUser(info.reaction.user);
-
-            if (target.count === 0) {
-                this.reactions = this.reactions.filter(r => r.reaction !== reaction);
-            }
-        }
-    }
-}
-
-interface SubmitContext {
-    channelLikeID: string;
-    dataStore: DataStore;
-    teamID: string;
-
-    extraInfo: string;
-    initialText: string;
-
-    emoji: EmojiService;
-
-    submit(text: string): Promise<any>;
-
-    changeChannelRequest(next: boolean);
-}
-
-class PostMessageContext implements SubmitContext {
-    constructor(
-        public client: SlackService,
-        public channelLikeID: string,
-        public teamID: string,
-        public infos: DisplaySlackMessageInfo[],
-    ) {
-    }
-
-    get dataStore(): DataStore {
-        return this.client.dataStore;
-    }
-
-    get emoji(): EmojiService {
-        return this.client.emoji;
-    }
-
-    get lastMessageTs(): string {
-        for (let i = 0; i < this.infos.length; i++) {
-            if (this.infos[i].message.channelID === this.channelLikeID) {
-                return this.infos[i].message.ts;
-            }
-        }
-        return '';
-    }
-
-    get initialText(): string {
-        return '';
-    }
-
-    get extraInfo(): string {
-        return '';
-    }
-
-    async submit(text: string): Promise<any> {
-        if (text.trim().match(/^\+:(.*):$/)) {
-            let reaction = text.trim().match(/^\+:(.*):$/)[1];
-            return this.client.addReaction(reaction, this.channelLikeID, this.lastMessageTs);
-        } else if (text.trim().match(/^\-:(.*):$/)) {
-            let reaction = text.trim().match(/^\-:(.*):$/)[1];
-            return this.client.removeReaction(reaction, this.channelLikeID, this.lastMessageTs);
-        } else {
-            return this.client.postMessage(this.channelLikeID, text);
-        }
-    }
-
-    changeChannelRequest(next: boolean) {
-        const channels: [string, SlackService, string][] = [];
-        for(const info of this.infos) {
-            if(!channels.find(c => c[0] === info.message.channelID)) {
-                channels.push([info.message.channelID, info.client, info.message.teamID]);
-            }
-        }
-
-        if(channels.length === 0) { return; }
-
-        const index = channels.findIndex(c => c[0] === this.channelLikeID);
-
-        const nextIndex = (index + (next ? 1 : -1) + channels.length) % channels.length;
-        this.channelLikeID = channels[nextIndex][0];
-        this.client = channels[nextIndex][1];
-        this.teamID = channels[nextIndex][2];
-    }
-}
-
-class EditMessageContext implements SubmitContext {
-    constructor(
-        public client: SlackService,
-        public message: SlackMessage,
-    ) {
-    }
-
-    get dataStore(): DataStore {
-        return this.client.dataStore;
-    }
-
-    get emoji(): EmojiService {
-        return this.client.emoji;
-    }
-
-    get channelLikeID(): string {
-        return this.message.channelID;
-    }
-
-    get channelName(): string {
-        return this.message.channelName;
-    }
-
-    get channelID(): string {
-        return this.message.channelID;
-    }
-
-    get teamID(): string {
-        return this.message.teamID;
-    }
-
-    get initialText(): string {
-        const messageText = this.message.text;
-        return messageText.replace(/<([^>]+)>/g, (value: string) => {
-            value = value.substr(1, value.length - 2);
-            const bar = value.indexOf('|');
-            if (bar >= 0) {
-                // Channel: <#XXYYZZ|channel>  =>  #channel
-                if(value[0] == '#')
-                    return '#' + value.substr(bar + 1);
-                // Url with no 'http': <http://github.com|github.com>  =>  github.com
-                else
-                    return value.substr(bar + 1);
-            } else {
-                // Full url: <https://github.com>  =>  https://github.com
-                return value;
-            }
-        }).replace(/&amp;/g, '&').replace(/&lt;/g, '<').replace(/&gt;/g, '>');
-    }
-
-    get extraInfo(): string {
-        return '(editing)';
-    }
-
-    async submit(text: string): Promise<any> {
-        return this.client.updateMessage(this.message.ts, this.message.channelID, text);
-    }
-
-    changeChannelRequest(next: boolean) {
-    }
-}
-
-interface FilterContext {
-    soloMode: boolean;
-    shouldShow(info: DisplaySlackMessageInfo): boolean;
-}
-
-class NoFilterContext implements FilterContext {
-    get soloMode(): boolean {
-        return false;
-    }
-
-    shouldShow(info: DisplaySlackMessageInfo): boolean {
-        return true;
-    }
-}
-
-class SoloChannelFilterContext implements FilterContext {
-    constructor(private channel: string) {
-    }
-
-    get soloMode(): boolean {
-        return true;
-    }
-
-    shouldShow(info: DisplaySlackMessageInfo): boolean {
-        return info.message.channelID === this.channel;
-    }
-}
-
+import { SubmitContext, PostMessageContext, EditMessageContext } from './submit-context';
+import { FilterContext, SoloChannelFilterContext, NoFilterContext } from './filter-context';
 
 @Component({
     selector: 'ss-list',
@@ -292,13 +35,16 @@ class SoloChannelFilterContext implements FilterContext {
     styles: [require('./slacklist.component.css').toString()],
 })
 export class SlackListComponent implements OnInit, OnDestroy {
-    messages: DisplaySlackMessageInfo[] = [];
-    slackServices: SlackService[];
+    slackServices: SlackClient[];
     submitContext: SubmitContext = null;
     filterContext: FilterContext = new NoFilterContext();
     subscription = new Subscription();
     submitting: boolean;
-    showingReactedUsers: number;
+    showingReactedUsers: any;
+
+    get messages(): DisplaySlackMessageInfo[] {
+        return this.slack.infos;
+    }
 
     get soloMode(): boolean {
         return this.filterContext.soloMode;
@@ -317,15 +63,14 @@ export class SlackListComponent implements OnInit, OnDestroy {
     }
 
     constructor(
-        private services: SlackServiceCollection,
+        private slack: SlackService,
         private events: GlobalEventService,
         private detector: ChangeDetectorRef,
         private router: Router,
         private setting: SettingService
     ) {
-        this.services.refresh();
-        this.slackServices = services.slacks;
-        this.messages = services.savedInfos;
+        this.slack.refresh();
+        this.slackServices = slack.clients;
     }
 
     ngOnInit(): void {
@@ -334,122 +79,21 @@ export class SlackListComponent implements OnInit, OnDestroy {
             return;
         }
 
-        for (const slack of this.slackServices) {
-            const parser = new ComposedParser([
-                new LinkParser(),
-                new NewLineParser(),
-                new EmojiParser(slack.emoji),
-                new MarkDownParser()
-            ]);
-
-            this.subscription.add(slack.messages.subscribe(message => this.onReceiveMessage(message, parser, slack)));
-            this.subscription.add(slack.reactionAdded.subscribe(reaction => this.onReactionAdded(reaction, parser, slack)));
-            this.subscription.add(slack.reactionRemoved.subscribe(reaction => this.onReactionRemoved(reaction, parser, slack)));
-            slack.start();
-        }
-
+        this.subscription.add(this.slack.onChange.subscribe(s => this.onChange(s)));
         this.subscription.add(this.events.activateMessageForm.subscribe(() => this.activateMessageForm()));
         this.subscription.add(this.events.keydown.filter(e => e.which === 38).subscribe(() => this.editLatestMessage()));
     }
 
     ngOnDestroy(): void {
-        for (const slack of this.slackServices) {
-            slack.stop();
-        }
-        this.services.savedInfos = this.messages;
         this.subscription.unsubscribe();
-    }
-
-    async onReactionAdded(reaction: SlackReactionAdded, parser: SlackParser, client: SlackService): Promise<void> {
-        const target = this.messages.find(m => m.message.rawMessage.ts === reaction.reaction.item.ts);
-        if (target) {
-            target.addReaction(reaction);
-        }
-        console.log(reaction.reaction);
-        this.detector.detectChanges();
-    }
-
-    async onReactionRemoved(reaction: SlackReactionAdded, parser: SlackParser, client: SlackService): Promise<void> {
-        const target = this.messages.find(m => m.message.rawMessage.ts === reaction.reaction.item.ts);
-        if (target) {
-            target.removeReaction(reaction);
-        }
-        console.log(reaction.reaction);
-        this.detector.detectChanges();
-    }
-
-    async onReceiveMessage(message: SlackMessage, parser: SlackParser, client: SlackService): Promise<void> {
-        console.log(message.rawMessage);
-
-        switch (message.rawMessage.subtype) {
-            case 'message_deleted':
-                await this.removeDeletedMessage(message, parser, client);
-                break;
-            case 'message_changed':
-                await this.changeMessage(message, parser, client);
-                break;
-            case 'message_replied':
-                await this.replyMessage(message, parser, client);
-                break;
-            default:
-                await this.addMessage(message, parser, client);
-                break;
-        }
-
-        this.detector.detectChanges();
-    }
-
-    async addMessage(message: SlackMessage, parser: SlackParser, client: SlackService): Promise<void> {
-        if (message.message) {
-            const info = new DisplaySlackMessageInfo(message, parser, client);
-            this.messages.unshift(info);
-
-            if (message.message.file && message.message.file.mimetype.indexOf('image') !== -1) {
-                info.image = await client.getImage(this.getMaximumThumbnail(message)).catch(e => {
-                    console.log('Getting image does not work in developing mode currently');
-                    return undefined;
-                });
-            }
-
-            client.markRead(message.channelID, message.ts);
-        }
-    }
-
-    getMaximumThumbnail(message: SlackMessage): string {
-        const file = message.rawMessage.file;
-        if (file.thumb_480) { return file.thumb_480; }
-        if (file.thumb_360) { return file.thumb_360; }
-        if (file.thumb_160) { return file.thumb_160; }
-        if (file.thumb_80) { return file.thumb_80; }
-        if (file.thumb_64) { return file.thumb_64; }
-        return '';
-    }
-
-    async deleteMessage(message: SlackMessage, client: SlackService): Promise<void> {
-        if (message.message) {
-            client.deleteMessage(message.channelID, message.ts);
-        }
-    }
-
-    async replyMessage(message: SlackMessage, parser: SlackParser, client: SlackService): Promise<void> {
-        // TODO
-    }
-
-    async removeDeletedMessage(message: SlackMessage, parser: SlackParser, client: SlackService): Promise<void> {
-        this.messages = this.messages.filter(m => message.rawMessage.deleted_ts !== m.message.rawMessage.ts);
-    }
-
-    async changeMessage(message: SlackMessage, parser: SlackParser, client: SlackService): Promise<void> {
-        const edited = this.messages.find(m => m.message.rawMessage.ts === message.rawMessage.message.ts);
-        if (edited) {
-            edited.edited = true;
-            edited.message.text = message.rawMessage.message.text;
-            edited.message.rawMessage.attachments = message.rawMessage.message.attachments;
-        }
     }
 
     get showForm(): boolean {
         return this.submitContext != null;
+    }
+
+    onChange(slack: SlackService): void {
+        this.detector.detectChanges();
     }
 
     onClickWrite(info: DisplaySlackMessageInfo) {
@@ -463,7 +107,7 @@ export class SlackListComponent implements OnInit, OnDestroy {
     }
 
     onClickDelete(info: DisplaySlackMessageInfo) {
-        this.deleteMessage(info.message, info.client);
+        this.slack.deleteMessage(info.message, info.client);
     }
 
     onClickSoloMode(info: DisplaySlackMessageInfo) {
